// src/app/models/producto/filtro-producto.ts
export interface FiltroProducto {
    idClr?: number[];   // Lista de números, por ejemplo, [1, 2, 3]
    idTll?: number[];   // Lista de números, por ejemplo, [1, 2, 3]
    idCtg?: number[];   // Lista de números, por ejemplo, [1, 2, 3]
    idMrc?: number[];   // Lista de números, por ejemplo, [1, 2, 3]
    idPrn?: number[];   // Lista de números, por ejemplo, [1, 2, 3]
    idMtl?: number[];   // Lista de números, por ejemplo, [1, 2, 3]
}